Data.NatMap

Documentation

data NatMap (f :: Type -> Type) (g :: Type -> Type)

Data structure which represents partial natural transformations, like usual Map which represents partial functions.

Map k v can be seen as a partial function.

m :: Map k v
(k -> Data.Map.lookup k m) :: k -> Maybe v

Analogically, a NatMap f g can be seen as a partial natural transformation, in the same sense Map represents a partial function.

nm :: NatMap f g
(fa -> lookup fa nm) :: forall a. f a -> Maybe (g a)

Entry

data Entry (f :: Type -> Type) (g :: Type -> Type)

getKeyValue :: forall (f :: Type -> Type) g. Entry f g -> (Shape f, g Var)

makeEntry :: (Traversable f, Traversable g, Ord k) => f k -> g k -> Maybe (Entry f g)

makeIdEntry :: Traversable f => f k -> Entry f f

unsafeMakeEntry :: (Traversable f, Functor g, Ord k) => f k -> g k -> Entry f g

Construction

empty :: forall (f :: Type -> Type) (g :: Type -> Type). NatMap f g

singleton :: forall (f :: Type -> Type) (g :: Type -> Type). Entry f g -> NatMap f g

partialIdentity :: (WeakOrd f, Traversable f) => [f a] -> NatMap f f

fromEntries :: forall (f :: Type -> Type) (g :: Type -> Type). WeakOrd f => [Entry f g] -> NatMap f g

insert :: forall (f :: Type -> Type) (g :: Type -> Type). WeakOrd f => Entry f g -> NatMap f g -> NatMap f g

delete :: forall f any (g :: Type -> Type). WeakOrd f => f any -> NatMap f g -> NatMap f g

Queries

size :: forall (f :: Type -> Type) (g :: Type -> Type). NatMap f g -> Int

member :: forall f a (g :: Type -> Type). WeakOrd f => f a -> NatMap f g -> Bool

notMember :: forall f a (g :: Type -> Type). WeakOrd f => f a -> NatMap f g -> Bool

lookup :: (WeakOrd f, Foldable f, Functor g) => f a -> NatMap f g -> Maybe (g a)

lookup_ :: forall (f :: Type -> Type) g. (WeakOrd f, Functor g) => Shape f -> NatMap f g -> Maybe (g Var)

keys :: forall (f :: Type -> Type) (g :: Type -> Type). NatMap f g -> [Shape f]

toEntries :: forall (f :: Type -> Type) (g :: Type -> Type). Traversable f => NatMap f g -> [Entry f g]

Map, Filter, Traversal

map1 :: forall g h (f :: Type -> Type). (forall a. g a -> h a) -> NatMap f g -> NatMap f h

mapMaybe1 :: forall g h (f :: Type -> Type). (forall a. g a -> Maybe (h a)) -> NatMap f g -> NatMap f h

traverse1 :: forall m g h (f :: Type -> Type). Applicative m => (forall a. g a -> m (h a)) -> NatMap f g -> m (NatMap f h)

wither1 :: forall m g h (f :: Type -> Type). Applicative m => (forall a. g a -> m (Maybe (h a))) -> NatMap f g -> m (NatMap f h)

mapWithKey1 :: Traversable f => (forall a. Ord a => f a -> g a -> h a) -> NatMap f g -> NatMap f h

mapMaybeWithKey1 :: Traversable f => (forall a. Ord a => f a -> g a -> Maybe (h a)) -> NatMap f g -> NatMap f h

traverseWithKey1 :: (Traversable f, Applicative m) => (forall a. Ord a => f a -> g a -> m (h a)) -> NatMap f g -> m (NatMap f h)

witherWithKey1 :: (Traversable f, Applicative m) => (forall a. Ord a => f a -> g a -> m (Maybe (h a))) -> NatMap f g -> m (NatMap f h)

Combinators

union :: forall (f :: Type -> Type) (g :: Type -> Type). WeakOrd f => NatMap f g -> NatMap f g -> NatMap f g

unionWith :: forall (f :: Type -> Type) g. WeakOrd f => (forall a. g a -> g a -> g a) -> NatMap f g -> NatMap f g -> NatMap f g

consistentUnion :: forall (f :: Type -> Type) (g :: Type -> Type). (WeakOrd f, Eq (g Var)) => NatMap f g -> NatMap f g -> Maybe (NatMap f g)

As partial natural transformation

identity :: forall (f :: Type -> Type). PTraversable f => NatMap f f

compose :: forall (f :: Type -> Type) (g :: Type -> Type) (h :: Type -> Type). (WeakOrd f, WeakOrd g, Foldable g, Functor h) => NatMap g h -> NatMap f g -> NatMap f h

outerNat :: forall (f :: Type -> Type) (g :: Type -> Type) (h :: Type -> Type). (Traversable f, Traversable g, PTraversable h, Ord (f (h Ignored))) => NatMap f g -> NatMap (Compose f h) (Compose g h)

innerNat :: forall (f :: Type -> Type) (g :: Type -> Type) (h :: Type -> Type). (Traversable f, Traversable g, PTraversable h, WeakOrd f) => NatMap f g -> NatMap (Compose h f) (Compose h g)

horizontalCompose :: forall (f :: Type -> Type) (g :: Type -> Type) (h :: Type -> Type) (j :: Type -> Type). (Traversable f, Functor g, Traversable h, Functor j, Ord (f (h Ignored))) => NatMap f g -> NatMap h j -> NatMap (Compose f h) (Compose g j)

Totality

fullSize :: forall (f :: Type -> Type) (g :: Type -> Type). PTraversable f => NatMap f g -> Int

isTotal :: forall (f :: Type -> Type) (g :: Type -> Type). PTraversable f => NatMap f g -> Bool

toTotal :: forall (f :: Type -> Type) (g :: Type -> Type). (PTraversable f, Functor g) => NatMap f g -> Maybe (f :~> g)

Re-exports

newtype (f :: k -> Type) :~> (g :: k -> Type) infixr 0

A natural transformation suitable for storing in a container.

Constructors

NT

Fields

($$) :: f ~> g

wrapNT :: (forall (a :: k). f a -> g a) -> f :~> g

wrapNT builds our natural transformation abstraction out of a natural transformation function.

An alias to NT provided for symmetry with unwrapNT.

unwrapNT :: Transformation f g t => t -> forall (a :: k). f a -> g a

unwrapNT is the nonfix version of #. It is used to break natural transformation wrappers, including :~>.

Utility

data Var

An opaque type representing syntactic variable.

indices :: Traversable f => Shape f -> f Var